add simpleRegularization function
simpleRegularization is both the standard and an example regularization algorithm.  It regularizes all nodes under the given input node such that all nodes are as specific as its descendant training reactions.  Ex:  If the group at a node has a [C,N] for a given atom for that node at least one descendant training reaction has a C for that atom and at least one descendant training reaction has an N for that atom.  Regularization operations are done by calling extendRegularization.  The algorithm starts at some node and goes through its descendant by recursively calling itself on the children of the node it was called on.